+++ src/index.ts
@@ -1,4 +1,0 @@
-import Graph, { GraphType as g } from './graph';
-
-export const TerraGraph = Graph;
-export type GraphType = g;